backend/requirements.txt
fastapi>=0.112.2
uvicorn>=0.29.0
SQLAlchemy>=2.0.0
alembic>=1.13.0
pydantic[email]>=2.0.0
pydantic-settings>=2.0.0
asyncpg>=0.29.0
redis[hiredis]>=5.0.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
httpx>=0.27.0
APScheduler>=3.10.0
pywebpush>=2.0.0
Pillow>=10.0.0
python-multipart>=0.0.9
boto3>=1.34.0
google-generativeai>=0.8.0
pytest>=8.0.0
pytest-asyncio>=0.23.0
